+++ src/Model/Command/UserRecommendation.php
@@ -31,14 +31,20 @@ class UserRecommendation extends AbstractCommand implements UserAwareInterface
     private $minimalRelevance = self::MINIMAL_RELEVANCE_LOW;
     /** @var array */
     private $filters = ['valid_to >= NOW'];
+    /** @var string|null */
+    private $modelName = null;
 
-    private function __construct(string $userId, int $count, string $scenario, float $rotationRate, int $rotationTime)
+    private function __construct(string $userId, int $count, string $scenario, float $rotationRate, int $rotationTime, string $modelName = null)
     {
         $this->setUserId($userId);
         $this->setCount($count);
         $this->setScenario($scenario);
         $this->setRotationRate($rotationRate);
         $this->setRotationTime($rotationTime);
+
+        if ($modelName !== null) {
+            $this->setModelName($modelName);
+        }
     }
 
     /**
@@ -51,6 +57,7 @@ class UserRecommendation extends AbstractCommand implements UserAwareInterface
      * Set from 0.0 for no rotation (same items will be recommended) up to 1.0 (same items should not be recommended).
      * @param int $rotationTime Specify for how long will the item's rotationRate be taken in account and so the item
      * is penalized for recommendations.
+     * @param string|null $modelName Specify which model you want to use
      * @return static
      */
     public static function create(
@@ -58,9 +65,10 @@ class UserRecommendation extends AbstractCommand implements UserAwareInterface
         int $count,
         string $scenario,
         float $rotationRate,
-        int $rotationTime
+        int $rotationTime,
+        string $modelName = null
     ): self {
-        return new static($userId, $count, $scenario, $rotationRate, $rotationTime);
+        return new static($userId, $count, $scenario, $rotationRate, $rotationTime, $modelName);
     }
 
     /**
@@ -121,6 +129,20 @@ class UserRecommendation extends AbstractCommand implements UserAwareInterface
         return $this;
     }
 
+    /***
+     * Set A/B model name
+     *
+     * @return $this
+     */
+    public function setModelName(string $modelName): self
+    {
+        Assertion::typeIdentifier($modelName);
+
+        $this->modelName = $modelName;
+
+        return $this;
+    }
+
     public function getUserId(): string
     {
         return $this->userId;
@@ -173,7 +195,7 @@ class UserRecommendation extends AbstractCommand implements UserAwareInterface
 
     protected function getCommandParameters(): array
     {
-        return [
+        $parameters = [
             'user_id' => $this->userId,
             'count' => $this->count,
             'scenario' => $this->scenario,
@@ -183,5 +205,11 @@ class UserRecommendation extends AbstractCommand implements UserAwareInterface
             'min_relevance' => $this->minimalRelevance,
             'filter' => $this->assembleFiltersString(),
         ];
+
+        if ($this->modelName !== null) {
+            $parameters['model_name'] = $this->modelName;
+        }
+
+        return $parameters;
     }
 }