+++ tests/integration/RequestBuilder/SortingRequestTest.php
@@ -2,6 +2,7 @@
 
 namespace Lmc\Matej\IntegrationTests\RequestBuilder;
 
+use Lmc\Matej\Exception\RequestException;
 use Lmc\Matej\IntegrationTests\IntegrationTestCase;
 use Lmc\Matej\Model\Command\Interaction;
 use Lmc\Matej\Model\Command\Sorting;
@@ -33,7 +34,7 @@ class SortingRequestTest extends IntegrationTestCase
     {
         $response = $this->createMatejInstance()
             ->request()
-            ->sorting(Sorting::create('user-b', ['item-a', 'item-b', 'itemC-c']))
+            ->sorting(Sorting::create('user-b', ['item-a', 'item-b', 'itemC-c'], 'default'))
             ->setUserMerge(UserMerge::mergeInto('user-b', 'user-a'))
             ->setInteraction(Interaction::bookmark('user-a', 'item-a'))
             ->send();
@@ -41,6 +42,19 @@ class SortingRequestTest extends IntegrationTestCase
         $this->assertInstanceOf(SortingResponse::class, $response);
         $this->assertResponseCommandStatuses($response, 'OK', 'OK', 'OK');
         $this->assertShorthandResponse($response, 'OK', 'OK', 'OK');
+    }
+
+    /** @test */
+    public function shouldFailOnInvalidModelName(): void
+    {
+        $this->expectException(RequestException::class);
+        $this->expectExceptionCode(400);
+        $this->expectExceptionMessage('BAD REQUEST');
+
+        $this->createMatejInstance()
+            ->request()
+            ->sorting(Sorting::create('user-b', ['item-a', 'item-b', 'itemC-c'], 'invalid-model-name'))
+            ->send();
     }
 
     private function assertShorthandResponse(SortingResponse $response, $interactionStatus, $userMergeStatus, $sortingStatus): void